Move webview shared code to workbench/common instead of workbench-api
For #141921
