+++ src/vs/workbench/common/webview.ts
